[notification,test] Change expectation - timescale
Given that the date pickers only specify date, the times will default to 12:00 AM.
At time of writing, the current expectation is that the application notifies the user of any upcoming notification on the same day or 1 day after, regardless of hours/minutes/seconds.
At time of writing, the application is expected to be local only, therefore timezones are not a concern.